+++ packages/service-postgres/src/lib/index.js
@@ -1,5 +1,4 @@
-import { INIT_SERVICE, START_SERVICE } from '@forrestjs/core';
-import * as hooks from './hooks';
+import * as targets from './targets';
 
 import { default as init } from './init';
 import { default as start } from './start';
@@ -11,18 +10,18 @@ export { default as start } from './start';
 export { default as query } from './query';
 export { getHandler, getModel, registerModel, resetModels } from './conn';
 
-export default ({ registerHook, registerAction, createHook }) => {
-  registerHook(hooks);
+export default ({ registerTargets, registerAction, createExtension }) => {
+  registerTargets(targets);
 
   registerAction({
-    hook: INIT_SERVICE,
-    name: hooks.SERVICE_NAME,
+    target: '$INIT_SERVICE',
+    name: '$SERVICE_NAME',
     trace: __filename,
     handler: async ({ getConfig, setContext, getContext }, ctx) => {
       const connections = getConfig('postgres.connections');
 
       // Let features hook into the existing connections configuration
-      createHook.sync(hooks.POSTGRES_BEFORE_INIT, { connections });
+      createExtension.sync('$POSTGRES_BEFORE_INIT', { connections });
 
       for (const options of connections) {
         // #18 apply defaults to the models list
@@ -31,11 +30,11 @@ export default ({ registerHook, registerAction, createHook }) => {
         }
 
         const connName = options.connectionName || 'default';
-        createHook.sync(`${hooks.POSTGRES_BEFORE_INIT}/${connName}`, {
+        createExtension.sync(`${targets.POSTGRES_BEFORE_INIT}/${connName}`, {
           options,
         });
         await init(options, ctx);
-        createHook.sync(`${hooks.POSTGRES_AFTER_INIT}/${connName}`, {
+        createExtension.sync(`${targets.POSTGRES_AFTER_INIT}/${connName}`, {
           options,
         });
       }
@@ -48,7 +47,7 @@ export default ({ registerHook, registerAction, createHook }) => {
         resetModels,
       });
 
-      createHook.sync(hooks.POSTGRES_AFTER_INIT, {
+      createExtension.sync('$POSTGRES_AFTER_INIT', {
         ...getContext('pg'),
         connections,
       });
@@ -56,31 +55,31 @@ export default ({ registerHook, registerAction, createHook }) => {
   });
 
   registerAction({
-    hook: START_SERVICE,
-    name: hooks.SERVICE_NAME,
+    target: '$START_SERVICE',
+    name: '$SERVICE_NAME',
     trace: __filename,
     handler: async ({ getConfig, getContext }, ctx) => {
       const connections = getConfig('postgres.connections');
 
-      await createHook.serie(hooks.POSTGRES_BEFORE_START, {
+      await createExtension.serie('$POSTGRES_BEFORE_START', {
         ...getContext('pg'),
         connections,
       });
 
       for (const options of connections) {
         const connName = options.connectionName || 'default';
-        createHook.serie(`${hooks.POSTGRES_BEFORE_START}/${connName}`, {
+        createExtension.serie(`${targets.POSTGRES_BEFORE_START}/${connName}`, {
           ...getContext('pg'),
           registerModel: (model) => options.models.push(model),
         });
         await start(options, ctx);
-        createHook.serie(`${hooks.POSTGRES_AFTER_START}/${connName}`, {
+        createExtension.serie(`${targets.POSTGRES_AFTER_START}/${connName}`, {
           ...getContext('pg'),
           handler: getHandler(connName),
         });
       }
 
-      await createHook.serie(hooks.POSTGRES_AFTER_START, {
+      await createExtension.serie('$POSTGRES_AFTER_START', {
         ...getContext('pg'),
         connections,
       });