+++ packages/service-postgres/src/lib/logger.js
@@ -1,0 +1,5 @@
+exports.logError = (...args) => console.error(...args);
+exports.logWarn = (...args) => console.warn(...args);
+exports.logInfo = (...args) => console.info(...args);
+exports.logVerbose = (...args) => console.log(...args);
+exports.logDebug = (...args) => console.log(...args);